medical-ai-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.20.0
pillow==10.0.1
python-multipart==0.0.6
numpy==1.26.0
opencv-python==4.8.1.78
pydantic==2.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.13.0